% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchprice.R
\name{catchprice}
\alias{catchprice}
\title{Catch prices}
\usage{
catchprice(data, price, species, year, region)
}
\arguments{
\item{data, }{dataframe to be cleaned}

\item{price, }{value or series of values}

\item{species, }{charater or strem of characters}

\item{year, }{value or series of values}

\item{region, }{charater or strem of characters}
}
\description{
Function to calculate price per catch of a given species,  year and region
}
\examples{
catchprice(catches, 2000, 'Cabrilla', c(1990, 1995), c('Guanacaste'))
}

