% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectcatches.R
\name{selectcatches}
\alias{selectcatches}
\title{Filters our catches data by species, region, and a range od years. Will provide summary data by 
of catches grouped by Region, species, and year}
\usage{
selectcatches(data, species, region, fxn, startyear = 1990, endyear = 2013)
}
\arguments{
\item{data}{A dataframe.}

\item{species}{A character or string of characters.}

\item{region}{A character or string of characters.}

\item{function}{to be applied to groups of Reion and year}
}
\value{
dataframe filtered by specicies and region and return of 'FUN'
}
\description{
Filters our catches data by species, region, and a range od years. Will provide summary data by 
of catches grouped by Region, species, and year
}
\examples{
select.catches(catches, c('Pargo', 'Pargo Seda'), 'Golfito', mean)
}

